# apps/file-server/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VFS Link File Server API",
    "version": "1.0.0",
    "description": "HTTP API for browsing and operating on VFS Link files, uploads, shares, thumbnails, and storage-key drift. Deployments may enable HTTP Basic authentication for the entire service."
  },
  "servers": [
    {
      "url": "./",
      "description": "Current deployment and base path"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
    "/api/status": {
      "get": {
        "tags": ["Status"],
        "summary": "Get storage status and aggregate statistics",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "description": "Current storage status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": ["Files"],
        "summary": "List direct children of a directory",
        "operationId": "listFiles",
        "parameters": [
          { "$ref": "#/components/parameters/PathQuery" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/OffsetQuery" }
        ],
        "responses": {
          "200": {
            "description": "Directory listing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FilesResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/files/move": {
      "post": {
        "tags": ["Files"],
        "summary": "Move files or directories",
        "description": "Small operations complete synchronously with 200. Large directory operations can be queued and return 202; poll the operation URL from the Location header.",
        "operationId": "moveFiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/EntriesResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/files/rename": {
      "post": {
        "tags": ["Files"],
        "summary": "Rename a file or directory",
        "description": "Small operations complete synchronously with 200. Large directory operations can be queued and return 202.",
        "operationId": "renameFile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenameRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/EntriesResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/files/trash": {
      "post": {
        "tags": ["Trash"],
        "summary": "Move files or directories to trash",
        "description": "Small operations complete synchronously with 200. Large directory operations can be queued and return 202.",
        "operationId": "trashFiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PathsRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/EntriesResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/trash": {
      "get": {
        "tags": ["Trash"],
        "summary": "List trash entries",
        "operationId": "listTrash",
        "responses": {
          "200": { "$ref": "#/components/responses/EntriesResult" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/trash/restore": {
      "post": {
        "tags": ["Trash"],
        "summary": "Restore trash entries",
        "description": "Small operations complete synchronously with 200. Large directory operations can be queued and return 202.",
        "operationId": "restoreTrash",
        "requestBody": { "$ref": "#/components/requestBodies/TrashIds" },
        "responses": {
          "200": { "$ref": "#/components/responses/EntriesResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/trash/delete": {
      "post": {
        "tags": ["Trash"],
        "summary": "Permanently delete selected trash entries",
        "description": "At least one trash ID is required. The deletion may return synchronously or as an asynchronous operation.",
        "operationId": "deleteTrash",
        "requestBody": { "$ref": "#/components/requestBodies/TrashIds" },
        "responses": {
          "200": { "$ref": "#/components/responses/DeletedResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/trash/empty": {
      "post": {
        "tags": ["Trash"],
        "summary": "Permanently delete all trash entries",
        "description": "The deletion may return synchronously or as an asynchronous operation.",
        "operationId": "emptyTrash",
        "responses": {
          "200": { "$ref": "#/components/responses/DeletedResult" },
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/tree": {
      "get": {
        "tags": ["Files"],
        "summary": "List the directory tree one level deep",
        "operationId": "getTree",
        "parameters": [
          { "$ref": "#/components/parameters/PathQuery" }
        ],
        "responses": {
          "200": {
            "description": "Directory node and its direct child directories",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TreeNode" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/download": {
      "get": {
        "tags": ["Files"],
        "summary": "Download a file",
        "operationId": "downloadFile",
        "parameters": [
          { "$ref": "#/components/parameters/RequiredPathQuery" },
          {
            "name": "disposition",
            "in": "query",
            "description": "Use inline to request browser display; every other value uses attachment.",
            "schema": {
              "type": "string",
              "enum": ["attachment", "inline"],
              "default": "attachment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes. Content-Type is inferred from the file extension and Content-Disposition includes the filename.",
            "headers": {
              "Content-Disposition": {
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/shares/drafts": {
      "post": {
        "tags": ["Shares"],
        "summary": "Create a share draft",
        "operationId": "createShareDraft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateShareDraftRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/ShareResult" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
...
    "/api/uploads": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Create an upload session",
        "description": "For local storage, uploadUrl points to this API's content endpoint. For GCS it is an opaque resumable session URI. Send each chunk with a client-generated Content-Range in addition to the fixed headers in the response.",
        "operationId": "createUpload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUploadRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload session created, including its upload capability",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/drift": {
      "get": {
        "tags": ["Drift"],
        "summary": "Inspect metadata-to-object-key drift",
        "description": "This endpoint remains readable when drift mutation is disabled or unavailable. With refresh=true it performs a storage scan; otherwise it reads the current snapshot. An unavailable or missing snapshot is represented by a 200 response with available=false or snapshotStatus=missing.",
        "operationId": "getDriftSnapshot",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "schema": { "type": "boolean", "default": false }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/DriftStatus" }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/DriftScope" }
          },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/OffsetQuery" }
        ],
        "responses": {
          "200": {
            "description": "Drift capability and snapshot state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DriftSnapshotResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Enabled only when the deployment configures Basic authentication credentials."
      }
    },
    "parameters": {
      "PathQuery": {
        "name": "path",
        "in": "query",
        "description": "Normalized logical path relative to the VFS root. Empty means root.",
        "schema": { "$ref": "#/components/schemas/LogicalPath" }
      },
...
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "schema": { "type": "string", "default": "" }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 }
      },
      "OffsetQuery": {
        "name": "offset",
        "in": "query",
        "schema": { "type": "integer", "minimum": 0, "default": 0 }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Basic authentication is enabled and credentials are missing or invalid",
        "headers": {
          "WWW-Authenticate": { "schema": { "type": "string" } }
        },
        "content": { "text/plain": { "schema": { "type": "string", "example": "Unauthorized" } } }
      },
...
      "MethodNotAllowed": {
        "description": "Method not allowed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "InternalServerError": {
        "description": "Internal server or storage failure",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
...
      "StatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["storageDriver", "storageRoot", "stats", "generatedAt"],
        "properties": {
          "storageDriver": { "type": "string" },
          "storageRoot": { "type": "string" },
          "stats": { "$ref": "#/components/schemas/Stats" },
          "generatedAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
    }
  }
}